+++ src/pages/AnalysisRuns/index.jsx
@@ -1,7 +1,8 @@
 import { InboxOutlined, UploadOutlined } from '@ant-design/icons';
-import { Button, DatePicker, Table, Upload } from 'antd';
+import { Button, DatePicker, Popconfirm, Table, Upload } from 'antd';
 import classNames from 'classnames';
 import TableFooter from 'components/layout/TableFooterLoader';
+import ActionInitiator from 'components/widgets/ActionInitiator';
 import ResultTag from 'components/widgets/ResultTag';
 import map from 'lodash.map';
 import mapValues from 'lodash.mapvalues';
@@ -124,7 +125,6 @@ const AnalysisRuns = () => {
           {record.title}
         </Link>
       ),
-      width: 150,
     },
     {
       title: 'Creation Date',
@@ -159,6 +159,7 @@ const AnalysisRuns = () => {
     {
       title: 'Created By',
       dataIndex: 'user',
+      render: (value) => <ActionInitiator initiator={value} />,
     },
     {
       title: 'Run Type',
@@ -167,14 +168,26 @@ const AnalysisRuns = () => {
     {
       title: 'Actions',
       render: (_, run) => (
-        <Button
-          onClick={() => onUploadClick(run.id)}
-          icon={<UploadOutlined />}
-          type="link"
-          disabled={run.status === constants.runStatuses.published}
+        <Popconfirm
+          title={`Are you sure you want to overwrite the results for the ${run.title} run?`}
+          onConfirm={() => onUploadClick(run.id)}
+          placement="topRight"
+          disabled={
+            run.status !== constants.runStatuses.analysis &&
+            run.status !== constants.runStatuses.review
+          }
         >
-          Upload result
-        </Button>
+          <Button
+            onClick={() =>
+              run.status === constants.runStatuses.qpcr && onUploadClick(run.id)
+            }
+            icon={<UploadOutlined />}
+            type="link"
+            disabled={run.status === constants.runStatuses.published}
+          >
+            Upload result
+          </Button>
+        </Popconfirm>
       ),
     },
   ];