+++ src/pages/IntakeReceiptLog/index.jsx
@@ -2,6 +2,7 @@ import { CommentOutlined, DownOutlined } from '@ant-design/icons';
 import { Button, Dropdown, Form, Menu, Popover, Space, Table, Tag } from 'antd';
 import classNames from 'classnames';
 import TableFooter from 'components/layout/TableFooterLoader';
+import ActionInitiator from 'components/widgets/ActionInitiator';
 import IntakeReceiptLogModal from 'components/widgets/Intake/IntakeReceiptLogModal';
 import isEmpty from 'lodash.isempty';
 import mapValues from 'lodash.mapvalues';
@@ -72,15 +73,9 @@ const IntakeReceiptLog = () => {
   );
 
   const scannerMenu = (logId) => (
-    <Menu onClick={startSession(logId)}>
+    <Menu onClick={startSession(logId)} className={styles.scannerMenu}>
       {scanners.items.map((item) => (
-        <Menu.Item
-          key={item.id}
-          disabled={!item.is_active || !item.is_online}
-          className={
-            (!item.is_active || !item.is_online) && styles.disabledScanner
-          }
-        >
+        <Menu.Item key={item.id} disabled={!item.is_active || !item.is_online}>
           <p style={{ margin: 0 }}>
             {item.scanner_id} - model:
             {item.model}{' '}
@@ -216,6 +211,7 @@ const IntakeReceiptLog = () => {
     {
       title: 'Logged By',
       dataIndex: 'logged_by',
+      render: (value) => <ActionInitiator initiator={value} />,
     },
     {
       title: 'Shipped On',