+++ src/pages/ScanSessions/index.jsx
@@ -12,6 +12,8 @@ import {
 } from 'antd';
 import classNames from 'classnames';
 import TableFooter from 'components/layout/TableFooterLoader';
+import ActionInitiator from 'components/widgets/ActionInitiator';
+import SearchTooltip from 'components/widgets/SearchTooltip';
 import debounce from 'lodash.debounce';
 import moment from 'moment-timezone';
 import React, {
@@ -83,13 +85,13 @@ const ScanSessions = () => {
     }, [filtersState.dates]);
   };
 
+  useFetching();
+
   const handleResetFilters = () => {
     return filtersDispatch({
       type: 'reset',
     });
   };
-
-  useFetching();
 
   const navigateToScan = useCallback(
     ({ sessionId, scanId }) => {
@@ -127,11 +129,11 @@ const ScanSessions = () => {
     {
       title: 'Scanned by',
       dataIndex: 'scanned_by',
-      render: (value) => value ?? '-',
+      render: (value) => <ActionInitiator initiator={value} />,
     },
   ];
 
-  const loadMore = useCallback(() => {
+  const loadMore = () => {
     const filteringParams = {
       limit: constants.scanSessions.itemsLoadingCount,
       offset,
@@ -152,7 +154,7 @@ const ScanSessions = () => {
         ...params,
       },
     });
-  }, [dispatch, filtersState, offset]);
+  };
 
   const sendQuery = useCallback(
     (query) => {
@@ -189,9 +191,7 @@ const ScanSessions = () => {
     };
   }, [delayedQuery]);
 
-  const onChangeSearch = (e) => {
-    const { target } = e;
-
+  const onChangeSearch = ({ target }) => {
     filtersDispatch({
       type: 'setValue',
       payload: {
@@ -319,9 +319,7 @@ const ScanSessions = () => {
             {scan.pool_id}
           </Link>
         ),
-        scan_time: scan.scan_timestamp
-          ? moment(scan.scan_timestamp).format(constants.dateTimeFormat)
-          : '-',
+        scan_time: scan.scan_timestamp,
         scan_name: poolName,
         pool_size: scan.tubes_count,
         rack_id: scan.rack_id,
@@ -375,6 +373,8 @@ const ScanSessions = () => {
         sorter: (a, b) =>
           moment(a.scan_time).valueOf() - moment(b.scan_time).valueOf(),
         sortDirections: ['ascend', 'descend', 'ascend'],
+        render: (value) =>
+          value ? moment(value).format(constants.dateTimeFormat) : '-',
       },
       { title: 'Scanner', dataIndex: 'scanner', key: 'scanner' },
       {
@@ -431,14 +431,23 @@ const ScanSessions = () => {
               xl={{ span: 6, offset: 10 }}
               xxl={{ span: 7 }}
             >
-              <Input
-                size="middle"
-                prefix={<SearchOutlined />}
-                placeholder="Search..."
-                value={filtersState.search}
-                allowClear
-                onChange={onChangeSearch}
-              />
+              <SearchTooltip
+                searchFields={[
+                  'session name',
+                  'company ID',
+                  'company name',
+                  'company name short',
+                ]}
+              >
+                <Input
+                  size="middle"
+                  prefix={<SearchOutlined />}
+                  placeholder="Search..."
+                  value={filtersState.search}
+                  allowClear
+                  onChange={onChangeSearch}
+                />
+              </SearchTooltip>
             </Col>
             <Col
               xs={{ span: 24 }}